requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
fakeredis
google-cloud-storage
opentelemetry-api==1.29.0
opentelemetry-exporter-google-cloud==0.17b0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-sdk==0.17b0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
